+++ srcs/MineClearance.Core/Services/GameTimer.cs
@@ -1,6 +1,5 @@
 using MineClearance.Core.Interfaces;
 using System;
-using System.ComponentModel;
 using System.Diagnostics;
 
 namespace MineClearance.Core.Services;
@@ -13,47 +12,20 @@ internal sealed class GameTimer : IGameTimer
     /// <summary>
     /// 高精度计时器字段, 用于计算已运行的时间
     /// </summary>
-    private readonly Stopwatch _stopwatch;
-
-    /// <inheritdoc/>
-    public event PropertyChangedEventHandler? PropertyChanged;
+    private readonly Stopwatch _stopwatch = new();
 
     /// <inheritdoc/>
     public DateTime? FirstStartTime { get; private set; }
 
     /// <inheritdoc/>
-    public TimeSpan Elapsed { get => field + _stopwatch.Elapsed; private set; }
+    public TimeSpan Elapsed { get => field + _stopwatch.Elapsed; private set; } = TimeSpan.Zero;
 
     /// <inheritdoc/>
-    public bool IsRunning => _stopwatch.IsRunning;
-
-    /// <summary>
-    /// 构造函数
-    /// </summary>
-    public GameTimer()
+    public void Initial(DateTime startTime, TimeSpan elapsed)
     {
-        Elapsed = TimeSpan.Zero;
-        _stopwatch = new();
-    }
-
-    /// <inheritdoc/>
-    public void Refresh()
-    {
-        // 触发 Elapsed 属性的 PropertyChanged 事件, 以便通知 UI 更新显示的时间
-        PropertyChanged?.Invoke(this, new(nameof(Elapsed)));
-    }
-
-    /// <inheritdoc/>
-    public void Pause()
-    {
-        // 如果计时器没有在运行, 则不需要暂停
-        if (!_stopwatch.IsRunning) { return; }
-
-        // 暂停高精度计时器
-        _stopwatch.Stop();
-
-        // 通知 IsRunning 属性已更改
-        PropertyChanged?.Invoke(this, new(nameof(IsRunning)));
+        Debug.Assert(!_stopwatch.IsRunning, "GameTimer should not be running when initializing.");
+        FirstStartTime = startTime;
+        Elapsed = elapsed;
     }
 
     /// <inheritdoc/>
@@ -65,31 +37,17 @@ internal sealed class GameTimer : IGameTimer
         // 更新第一次开始计时的时间
         FirstStartTime ??= DateTime.Now;
 
-        // 启动高精度计时器
+        // 启动计时器
         _stopwatch.Start();
-
-        // 通知 IsRunning 属性已更改
-        PropertyChanged?.Invoke(this, new(nameof(IsRunning)));
     }
 
     /// <inheritdoc/>
-    public void ReStart()
+    public void Pause()
     {
-        // 更新第一次开始计时的时间
-        FirstStartTime ??= DateTime.Now;
+        // 如果计时器没有在运行, 则不需要暂停
+        if (!_stopwatch.IsRunning) { return; }
 
-        // 重置高精度计时器并重新开始计时
-        _stopwatch.Restart();
-
-        // 通知 IsRunning 属性已更改
-        PropertyChanged?.Invoke(this, new(nameof(IsRunning)));
-    }
-
-    /// <inheritdoc/>
-    public void Initial(DateTime startTime, TimeSpan elapsed)
-    {
-        Debug.Assert(!_stopwatch.IsRunning, "GameTimer should not be running when initializing.");
-        FirstStartTime = startTime;
-        Elapsed = elapsed;
+        // 暂停高精度计时器
+        _stopwatch.Stop();
     }
 }